minor:封装batch_request到client中
