+++ pipeline_plugins/variables/collections/sites/open/cmdb/var_set_module_selector.py
@@ -16,7 +16,7 @@ import logging
 from gcloud.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
-from gcloud.utils.cmdb import batch_request
+from api.utils.cmdb import batch_request
 from pipeline.core.data.var import LazyVariable
